Progress on module CSS and JS generation
Completed a python script to create a module's javascript file based on the javascript files of modules 1-6, and have finished roughly half of the script for generating a modules CSS files based on modules 1-6. One thing of note is that the javascript and css files for modules 1-6 were all identical, so there is little purpose to this as of right now.