Completion of ModCssGen script and overall modGen script
Completed the simple python scripts that create a module's css file, and another script that calls all three of the module generation scripts. Two issues remain: The first issue is getting the files for modHtml generation. The four text files for modHtml must follow the same naming scheme. The second issue is getting the video url to modHtmlGen as I don't necessarily want it to be an input parameter for modGen, but it does need to be given in some way.